+++ src/Tar/Encode.php
@@ -8,7 +8,7 @@ use Innmind\Filesystem\{
     File\Content,
     Directory,
 };
-use Innmind\TimeContinuum\{
+use Innmind\Time\{
     Clock,
     Format,
 };
@@ -32,6 +32,7 @@ final class Encode
         $this->clock = $clock;
     }
 
+    #[\NoDiscard]
     public function __invoke(File|Directory $file): Content
     {
         return Content::ofChunks(
@@ -50,6 +51,7 @@ final class Encode
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     public static function of(Clock $clock): self
     {
         return new self($clock);