+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Requires PHP `8.4`
+- Requires `innmind/filesystem:~9.0`
 
 ## 2.0.0 - 2025-04-16