+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 3.0.0 - 2026-02-01
+
+### Changed
+
+- Requires PHP `8.4`
+- Requires `innmind/filesystem:~9.0`
+- Requires `innmind/time:~1.0`
+
 ## 2.0.0 - 2025-04-16
 
 ### Changed